extract logic into new ModalCloseButton component
