+++ packages/adapters/src/Modal/Modal/styles.scss
