+++ packages/adapters/src/Modal/ModalCloseButton/styles.scss
