+++ packages/adapters/src/Modal/index.ts
@@ -1,4 +1,5 @@
 export { default as Modal } from './Modal';
-export { default as modalCloseButtonProps } from './modalCloseButtonProps';
+export { default as ModalCloseButton } from './ModalCloseButton';
 
-export * from './types';
+export * from './Modal/types';
+export * from './ModalCloseButton/types';
